requirements.txt
pika==1.3.2
python-dotenv==1.0.1
python3-nmap==1.9.1
simplejson==3.19.3
